requirements.txt
pytdbot~=0.8.6
pyyaml~=6.0
#lz4~=4.3.2